requirements.txt
pyjwt~=2.8.0
pyjwt[crypto]~=2.8.0
requests~=2.32.2